examples/mnist/requirements.txt
pytorch-ignite
tensorboard
